Add footer, router, and various pages with forms
This commit includes the addition of several new files such as Footer, Router, and various other pages with forms such as RegisterPage and LoginPage. It also includes addition of new dependencies for creating these pages and updating the package.json file accordingly. The newly added pages are incorporated into the routing system. The styling for these pages and components have also been introduced.