Introduce Material UI theme and improve social login buttons
Added the Material UI library theme with primary and secondary colors. The theme is utilized in the social login buttons on the LoginPage. Styling using emotion is added to the buttons to use the theme colors. Additionally, unnecessary comments are removed from the routing file while a localhost proxy is added to the package.json file.