Remove loading and status state from auth, redirect unauthenticated users at cart
The 'loading' and 'status' states have been removed from the auth component as they are unnecessary. Code commenting has been used to effect these removals. The ProductDetailsPage.jsx now checks if a user is authenticated before they can add an item to the cart. If they are not authenticated, they will be redirected to the login page.